# docs/swagger.yaml
basePath: /
definitions:
  models.Location:
    properties:
      coordinates:
        description: Array com [longitude, latitude]
        items:
          type: number
        type: array
      type:
        description: Será sempre "Point"
        type: string
    type: object
  models.Scan:
    properties:
      id:
        type: string
      location:
        $ref: '#/definitions/models.Location'
      qrcodeId:
        type: string
      scanedAt:
        type: string
    type: object
  qrcode.CreateQRCodeDto:
    properties:
      lat:
        type: number
      link:
        type: string
      long:
        type: number
      slug:
        maxLength: 20
        minLength: 2
        type: string
      userId:
        type: string
    required:
    - lat
    - link
    - long
    - slug
    - userId
    type: object
  qrcode.QRCodeWithURL:
    properties:
      createdAt:
        type: string
      id:
        type: string
      link:
        type: string
      location:
        $ref: '#/definitions/models.Location'
      slug:
        type: string
      updatedAt:
        type: string
      url:
        type: string
      userId:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is an API for managing users and QR codes.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: QR Code Boost API
  version: "1.0"
paths:
  /qr:
    post:
      consumes:
      - application/json
      parameters:
      - description: QR Code Payload
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/qrcode.CreateQRCodeDto'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/qrcode.QRCodeWithURL'
      summary: Create a QR Code
      tags:
      - QR Codes
  /qr/{slug}:
    get:
      consumes:
      - application/json
      parameters:
      - description: QR Code Slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/qrcode.QRCodeWithURL'
      summary: Find QR Code by Slug
      tags:
      - QR Codes
  /qr/near/{slug}:
    get:
      consumes:
      - application/json
      parameters:
      - description: QR Code Slug
        in: path
        name: slug
        required: true
        type: string
      - description: 'Maximum distance in meters (default: 3000 or 3km)'
        in: query
        name: maxDistance
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Scan'
            type: array
      summary: Find scans near a QR Code
      tags:
      - QR Codes
  /qr/user/{userId}:
    get:
      consumes:
      - application/json
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/qrcode.QRCodeWithURL'
            type: array
      summary: List QR Codes from a specific user
      tags:
      - QR Codes
swagger: "2.0"